IGNITE-28875 Refactor system view tests
