+++ modules/core/src/test/java/org/apache/ignite/testsuites/IgniteCacheTestSuite13.java
@@ -35,10 +35,25 @@ import org.apache.ignite.internal.metric.MetricsSelfTest;
 import org.apache.ignite.internal.metric.OutboundIoMessageQueueSizeTest;
 import org.apache.ignite.internal.metric.ReadMetricsOnNodeStartupTest;
 import org.apache.ignite.internal.metric.SystemMetricsTest;
+import org.apache.ignite.internal.metric.SystemViewBinaryMetaTest;
 import org.apache.ignite.internal.metric.SystemViewCacheExpiryPolicyTest;
+import org.apache.ignite.internal.metric.SystemViewCacheTest;
+import org.apache.ignite.internal.metric.SystemViewClientTest;
 import org.apache.ignite.internal.metric.SystemViewClusterActivationTest;
 import org.apache.ignite.internal.metric.SystemViewComputeJobTest;
-import org.apache.ignite.internal.metric.SystemViewSelfTest;
+import org.apache.ignite.internal.metric.SystemViewComputeTaskTest;
+import org.apache.ignite.internal.metric.SystemViewConfigurationTest;
+import org.apache.ignite.internal.metric.SystemViewDSTest;
+import org.apache.ignite.internal.metric.SystemViewExecutorsTest;
+import org.apache.ignite.internal.metric.SystemViewMetastorageTest;
+import org.apache.ignite.internal.metric.SystemViewNodesTest;
+import org.apache.ignite.internal.metric.SystemViewPageListsTest;
+import org.apache.ignite.internal.metric.SystemViewPageTimestampsTest;
+import org.apache.ignite.internal.metric.SystemViewPluginTest;
+import org.apache.ignite.internal.metric.SystemViewQueriesTest;
+import org.apache.ignite.internal.metric.SystemViewServiceTest;
+import org.apache.ignite.internal.metric.SystemViewSnapshotsTest;
+import org.apache.ignite.internal.metric.SystemViewTransactionsTest;
 import org.apache.ignite.internal.processors.cache.CacheClearAsyncDeadlockTest;
 import org.apache.ignite.internal.processors.cache.CacheDistributedGetLongRunningFutureDumpTest;
 import org.apache.ignite.internal.processors.cache.EntriesRemoveOnShutdownTest;
@@ -88,10 +103,25 @@ public class IgniteCacheTestSuite13 {
         GridTestUtils.addTestIfNeeded(suite, SystemMetricsTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, CustomMetricsTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, MetricsConfigurationTest.class, ignoredTests);
-        GridTestUtils.addTestIfNeeded(suite, SystemViewSelfTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, SystemViewClusterActivationTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, SystemViewComputeJobTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, SystemViewCacheExpiryPolicyTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewCacheTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewServiceTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewComputeTaskTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewClientTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewNodesTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewTransactionsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewDSTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewExecutorsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewPageListsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewBinaryMetaTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewMetastorageTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewSnapshotsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewPageTimestampsTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewConfigurationTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewPluginTest.class, ignoredTests);
+        GridTestUtils.addTestIfNeeded(suite, SystemViewQueriesTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, CacheMetricsAddRemoveTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, CacheMetricsConflictResolverTest.class, ignoredTests);
         GridTestUtils.addTestIfNeeded(suite, JmxExporterSpiTest.class, ignoredTests);
